Add asset import to timeline context menu
