+++ packages/studio/src/components/Canvas.tsx
@@ -1,5 +1,4 @@
 import type {Size} from '@remotion/player';
-import type {InsertableCompositionElement} from '@remotion/studio-shared';
 import React, {
 	useCallback,
 	useContext,
@@ -10,12 +9,8 @@ import React, {
 } from 'react';
 import type {CanvasContent} from 'remotion';
 import {Internals, watchStaticFile} from 'remotion';
-import {getStaticFiles} from '../api/get-static-files';
-import {writeStaticFile} from '../api/write-static-file';
 import {StudioServerConnectionCtx} from '../helpers/client-id';
 import {BACKGROUND} from '../helpers/colors';
-import {detectFileType} from '../helpers/detect-file-type';
-import type {FileType} from '../helpers/detect-file-type';
 import type {AssetMetadata} from '../helpers/get-asset-metadata';
 import {getAssetMetadata} from '../helpers/get-asset-metadata';
 import {
@@ -33,12 +28,11 @@ import {useKeybinding} from '../helpers/use-keybinding';
 import {canvasRef} from '../state/canvas-ref';
 import {EditorShowGuidesContext} from '../state/editor-guides';
 import {EditorZoomGesturesContext} from '../state/editor-zoom-gestures';
-import {callApi} from './call-api';
 import EditorGuides from './EditorGuides';
 import {EditorRulers} from './EditorRuler';
 import {useIsRulerVisible} from './EditorRuler/use-is-ruler-visible';
+import {importAssets} from './import-assets';
 import {SPACING_UNIT} from './layout';
-import {showNotification} from './Notifications/NotificationCenter';
 import {VideoPreview} from './Preview';
 import {ResetZoomButton} from './ResetZoomButton';
 import {useResolvedStack} from './Timeline/use-resolved-stack';
@@ -61,69 +55,6 @@ const resetZoom: React.CSSProperties = {
 };
 
 const ZOOM_PX_FACTOR = 0.003;
-
-const getAssetElement = ({
-	fileType,
-	src,
-}: {
-	fileType: FileType;
-	src: string;
-}): InsertableCompositionElement | null => {
-	if (
-		fileType.type === 'png' ||
-		fileType.type === 'jpeg' ||
-		fileType.type === 'webp' ||
-		fileType.type === 'bmp'
-	) {
-		return {
-			type: 'asset',
-			assetType: 'image',
-			src,
-			dimensions: fileType.dimensions,
-		};
-	}
-
-	if (fileType.type === 'gif') {
-		return {
-			type: 'asset',
-			assetType: 'gif',
-			src,
-			dimensions: fileType.dimensions,
-		};
-	}
-
-	if (
-		fileType.type === 'riff' ||
-		fileType.type === 'webm' ||
-		fileType.type === 'iso-base-media' ||
-		fileType.type === 'transport-stream'
-	) {
-		return {
-			type: 'asset',
-			assetType: 'video',
-			src,
-			dimensions: null,
-		};
-	}
-
-	return null;
-};
-
-const getAssetLabel = (element: InsertableCompositionElement) => {
-	if (element.type !== 'asset') {
-		throw new Error('Expected asset element');
-	}
-
-	if (element.assetType === 'image') {
-		return '<Img>';
-	}
-
-	if (element.assetType === 'video') {
-		return '<Video>';
-	}
-
-	return '<Gif>';
-};
 
 type WebKitGestureEvent = UIEvent & {
 	scale: number;
@@ -701,114 +632,13 @@ export const Canvas: React.FC<{
 				return;
 			}
 
-			const staticFiles = getStaticFiles();
-			const differentExistingFile = files.find((file) => {
-				return staticFiles.some(
-					(staticFile) =>
-						staticFile.name === file.name &&
-						staticFile.sizeInBytes !== file.size,
-				);
-			});
-			if (differentExistingFile) {
-				showNotification(
-					`File with name ${differentExistingFile.name} already exists and is different`,
-					4000,
-				);
-				return;
-			}
-
 			setIsAddingAsset(true);
-			const insertedLabels: string[] = [];
-			const addedStaticFiles: string[] = [];
-			const unsupportedFiles: string[] = [];
-			const notifyAddedStaticFiles = () => {
-				if (addedStaticFiles.length === 1) {
-					showNotification(
-						`Created ${addedStaticFiles[0]} in public folder`,
-						3000,
-					);
-				} else if (addedStaticFiles.length > 1) {
-					showNotification(
-						`Added ${addedStaticFiles.length} files to public folder`,
-						3000,
-					);
-				}
-
-				addedStaticFiles.length = 0;
-			};
-
 			try {
-				for (const file of files) {
-					const contents = await file.arrayBuffer();
-					const fileType = detectFileType(new Uint8Array(contents));
-					const element = getAssetElement({
-						fileType,
-						src: file.name,
-					});
-
-					if (element === null) {
-						unsupportedFiles.push(file.name);
-						continue;
-					}
-
-					const alreadyExists = staticFiles.some(
-						(staticFile) =>
-							staticFile.name === file.name &&
-							staticFile.sizeInBytes === file.size,
-					);
-
-					if (!alreadyExists) {
-						await writeStaticFile({
-							contents,
-							filePath: file.name,
-						});
-						addedStaticFiles.push(file.name);
-					}
-
-					const result = await callApi('/api/insert-jsx-element', {
-						compositionFile,
-						compositionId: currentCompositionId,
-						element,
-					});
-
-					if (!result.success) {
-						notifyAddedStaticFiles();
-						showNotification(result.reason, 4000);
-						return;
-					}
-
-					insertedLabels.push(getAssetLabel(element));
-				}
-
-				notifyAddedStaticFiles();
-
-				if (insertedLabels.length === 1) {
-					showNotification(`Added ${insertedLabels[0]} to source file`, 2000);
-				} else if (insertedLabels.length > 1) {
-					showNotification(
-						`Added ${insertedLabels.length} assets to source file`,
-						2000,
-					);
-				}
-
-				if (unsupportedFiles.length === 1) {
-					showNotification(
-						`Cannot add ${unsupportedFiles[0]}: Unsupported file type`,
-						3000,
-					);
-				} else if (unsupportedFiles.length > 1) {
-					showNotification(
-						`Skipped ${unsupportedFiles.length} unsupported files`,
-						3000,
-					);
-				}
-			} catch (error) {
-				showNotification(
-					`Could not add asset: ${
-						error instanceof Error ? error.message : String(error)
-					}`,
-					4000,
-				);
+				await importAssets({
+					files,
+					compositionFile,
+					compositionId: currentCompositionId,
+				});
 			} finally {
 				setIsAddingAsset(false);
 			}
